Moved embed stuff to it's own cog.
